Fix trust_remote_code in snippets (#467)
At the moment, we only add `trust_remote_code` in the cases where
there's a custom class, but this will fail when there are other custom
components, such as tokenizer/config/etc. Instead, we just use
`trust_remote_code` if there's an `auto_map` in the `config`, which is
the same condition to add the `custom_code` tag to repos. This fixes:

* https://huggingface.co/allenai/OLMo-7B
* https://huggingface.co/togethercomputer/m2-bert-80M-2k-retrieval

---------

Co-authored-by: Julien Chaumond <julien@huggingface.co>