gguf: Add ability to load local file (#656)
Being able to load a local gguf file can be useful when we want to debug
a gguf file.

**Without this PR**, this ability could be done by using
[file-fetch](https://www.npmjs.com/package/file-fetch). However, that
won't work with big models, since the whole file is loaded into RAM.

This PR add a new `RangeViewLocalFile` internal class that extends
`RangeView`. It redirects calls to `fetchChunk()` to
`fs.createReadStream` with the appropriate byte range. This allows the
library to read specific chunk from a local file.

For security reason, this ability is locked under `localFile: boolean`
param. By default, it is disabled (i.e. when this library is run on hub
backend, this param is disabled if unspecified)

- [x] Add test case to `gguf.spec.ts`
- [x] Being able to build with target=browser (only build, but will
throw error on browser if being used)