Add Sanctum local app to "Use this model" list (#702)
Hi HuggingFace team!

Sanctum is a local encrypted application for running open-source LLMs in
just a few clicks. It has a direct integration with HuggingFace GGUF
models, so it'd be great to have it under the list of "Use this model"
local apps. Sanctum is free and available for download on MacOS &
Windows here: [https://sanctum.ai](https://sanctum.ai)

We've also implemented support for deep links, here's a quick video on
how it works on MacOS (Windows is also supported). Feel free to use this
link to try it out:
`sanctum://open_from_hf?model=SanctumAI/Meta-Llama-3-8B-Instruct-GGUF`


https://github.com/huggingface/huggingface.js/assets/32494918/9f9acc09-0396-48b7-ba8b-9945c6dd547c

Also is there anything we could do for Sanctum to be available in the
"Use this model" app list by default?

Thanks in advance!

Co-authored-by: Bertrand CHEVRIER <bertrand@huggingface.co>