PR template when adding a new library (#913)
This PR adds a template with a checklist to follow when adding a new
library.

Based on
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository.

Once merged, we should be able to generate a link to create such a PR
(the template will not be used by default for all PRs :smile:). This
link would be added to
https://huggingface.co/docs/hub/models-download-stats#can-i-add-my-query-files-for-my-library.

Hopefully, this should reduce the back-and-fourth when someone adds a
new library. I suggest we start with that as an experiment and if it
doesn't work, we can move it somewhere else as documentation.