Update TGI specs (+ define TextToSpeech independently) (#915)
This PR pulls latest changes from TGI specs. In particular:
- updated `ChatCompletionInputMessageContent` => now supports single
message (as before), array of messages and **sending image as input**
(in case of VLMs)
- new input: `stream_options`
- tools input have been updated
- more documented fields

**Note:** TGI specs have a `guideline` input parameter for chat
completion. I removed it from the official specs as it's a very specific
parameter and subject to deletion in the near future (it was there to
fix a chat template issue for `ShieldGemma` models). More context
[here](https://huggingface.slack.com/archives/C05CFK1HM0T/p1726822675255589?thread_ts=1724852501.099809&cid=C05CFK1HM0T)
(internal).

**Note:** I also took the opportunity to define `TextToSpeech` specs
independently from `TextToAudio` (otherwise the semi-automatic scripts
are not working properly). Nothing has changed in the specs.

I'll open a related PR on `huggingface_hub` side to reflect this change.