[integration] Add `hezar` to libraries and snippets (#934)
This PR adds integration for `hezar` mainly for download stats and
supporting code snippets for the models built with this library. This
library currently has 30 models and 10 datasets with 70K installs and
counting. This library has been around for a year now and since our
users requested this a lot, we decided to add this integration for
convenience.

Library repo: https://github.com/hezarai/hezar
Creator: @arxyzan

Let me know if this PR conforms to the necessary guidelines. Thank you
guys ;)