Added CoTracker for download stats (#925)
Hey! 

Thanks for the amazing library :) 

We are hosting CoTracker model checkpoint
[here](https://huggingface.co/facebook/cotracker/tree/main) and would
like to see the download count. This PR follows [this
one](https://github.com/huggingface/huggingface.js/pull/561/files) to
enable it.

Hope it's correct way to do it, let me know if any changes are needed or
there is a better way to do so.

---------

Co-authored-by: Lucain <lucainp@gmail.com>