Fix zero-shot tasks specs (#1037)
Related to [slack
thread](https://huggingface.slack.com/archives/C069KSP486B/p1731923978189269)
(private).

This PR fixes the specs for `zero-shot-classification`,
`zero-shot-image-classification` and `zero-shot-object-detection`. The
candidate labels must be sent as parameters, not in `inputs`. The
Inference API itself **did not change**, it's just the specs that were
not aligned with the existing API.