Refacto Inference snippets tests (autogeneration + 1 snippet == 1 file) (#1046)
This PR is a refacto of the inference snippet tests. The main goal is to
reduce friction when updating the inference snippets logic. In
particular:
- each generated snippet is saved into a single test file with the
correct file extension => should greatly improve reviews. Currently
snippets are saved as hardcoded strings in JS files, making them hard to
review (no syntax highlighting, weird indentation).
- snippets can be updated with a script, reducing the need to manually
edit tests (which is quite time-consuming)

This PR is quite large given the auto-generated files. The main parts to
review are:
- `generate-snippets-fixtures.ts` => the script that generates the
snippets and test them.
- package.json, pnpm lock files, etc. => I'm not entirely sure of what I
did there (especially to make `tasks-gen` depend on `tasks`)
- `python.spec.ts` / `js.specs.ts` and `curl.specs.ts` have been removed
- everything in `packages/tasks-gen/snippets-fixtures/` => the test
cases. I've only committed the ones that where previously tested.

Thanks to this PR, I fixed a typo in the JS snippets (a missing comma)
and curl snippets (consistency between `"` and `'`).


cc @mishig25 with whom I quickly discussed this

---------

Co-authored-by: Mishig <dmishig@gmail.com>