Add CXR Foundation and Derm Foundation model libraries (Google Health AI) (#1049)
For two Google Health AI models, a primary usage pattern (with
supporting documentation and notebooks) is downloading only the
precomputed .npz embeddings. As such, I'm submitting for two new model
libraries which are essentially an extension of the TF Keras model
library, but they cleanly handle the non-`saved_model.pb` usage pattern
in the `countDownload` override. This PR was designed to conform with
https://huggingface.co/docs/hub/en/models-download-stats and
https://huggingface.co/docs/hub/en/models-adding-libraries#register-your-library.

Public models https://huggingface.co/google/cxr-foundation and
https://huggingface.co/google/derm-foundation will be switched over once
this PR is merged.

---------

Co-authored-by: vb <vaibhavs10@gmail.com>