Generate KerasHub snippets based on tasks from `metadata.json` (#1118)
This PR updates the `keras-hub` snippets based on the new `metadata.json
> tasks` field. This field is now uploaded for all KerasHub models (see
https://github.com/keras-team/keras-hub/pull/1997) and contains the list
of tasks compatible for a given model. This allows us to generate
multiple snippets when relevant. For instance,
[keras/stable_diffusion_3.5_large_turbo](https://huggingface.co/keras/stable_diffusion_3.5_large_turbo/blob/main/metadata.json)
is compatible with `ImageToImage`, `Inpaint` and `TextToImage` tasks.

For this PR to work, we'll need to parse the metadata.json file
server-side. This is done in
https://github.com/huggingface-internal/moon-landing/pull/11693 (private
PR). We can merge these 2 PRs independently.

cc @martin-gorner @mattdangerw @SamanehSaadat who coordinated this 

---

**Note:** I also removed the legacy `keras-nlp` library (only [18
remaining models](https://huggingface.co/models?library=keras-nlp)).