Accept string[] as feature-extraction input (#1166)
Related to https://github.com/huggingface/huggingface_hub/issues/2824.

This PR makes it possible to send a `string[]` instead of `string` as
`feature-extraction` inputs. This is already possible in practice in
Inference API but not documented.

In the past, I've pushed back on this change (see
https://github.com/huggingface/huggingface_hub/issues/1745 and
https://github.com/huggingface/huggingface_hub/pull/1746#issuecomment-1766640525)
but I think it's fine to revisit it now. The main reason I mentioned was
that `feature-extraction`'s server-side implementation was mostly a
for-loop on the text input so acception a `string[]` would not really
improve performances. That been said, there has been quite some
improvements since then and especially the `text-embedding-inference`
framework.