[draft] Support providerModelId in inference snippets (#1210)
Goal of this PR is to correctly handle model id mapping when generating
inference snippets for a given provider.

For now it's a simple PR to showcase what I had in mind in
https://github.com/huggingface-internal/moon-landing/pull/12626#discussion_r1961923510
(private repo). I only implemented it for chat-completion curl snippet
but the rest would follow the same pattern.

Note that we should use the mapping "only" for curl, pure Python, pure
JS and openai client snippets. For `huggingface.js` and
`huggingface_hub` ones, the model id from the Hub should be used.


---

**Note:** orthogonal to this PR but I realized that some URLs are also
incorrect depending on the provider.

---------

Co-authored-by: SBrandeis <simon@huggingface.co>