Adding Merlin for accurate download count (#1232)
I don't think we are getting an accurate download count for our hugging
face repo: https://huggingface.co/stanfordmimi/Merlin, since the
downloads page counts the queries to the config.json file and in our
case the config.json file is empty. I was looking through the hugging
face documentation and it seems like I had to add the file to the
hugging face repo to allow it to have an accurate download count? Please
let me know if I'm mistaken.

---------

Co-authored-by: Lucain <lucainp@gmail.com>
Co-authored-by: Lucain <lucain@huggingface.co>