Do not require all dummy api keys in inference tests (#1226)
I currently find it annoying when I want to run tests locally (that
already have cached tapes) to set the env variables to `"dummy"`:

```sh
HF_FAL_KEY=dummy  pnpm run test -t "Fal AI"
```

This PR is a suggestion to get rid of that and default to `"dummy"` in
the code instead of in the CI workflow file. Happy to get better
suggestions if you think of one (in Python I would do an automatic
fixture "get_api_key_or_dummy" but quite specific to pytest).

As a bonus, no need to update `.github/workflows/test.yml` when adding a
new provider.

WDYT?