[Inference Providers] Async calls for `text-to-video` with fal.ai (#1292)
## What does this PR do?

This PR adds asynchronous polling to the fal.ai text-to-video
generation. This allows running inference with models that may take > 2
min to generate results. The other motivation behind this PR is to align
the Python and JS clients, the Python client has already been merged
into main: https://github.com/huggingface/huggingface_hub/pull/2927


## Main Changes
- Replaced static `baseUrl` property with `makeBaseUrl()` function
across all providers. This is needed to be able to customize the base
url based on the task. We want to use `FAL_AI_API_BASE_URL_QUEUE` for
`text-to-video` only. I'm not convinced if it's the simplest and the
best way to do that.
- Added a `pollFalResponse()` for `text-to-video`(similarly to what it's
done with BFL for `text-to-image`).

Any refactoring suggestions are welcome! I'm willing to spend some
additional time to make provider-specific updates easier to implement
and better align our two clients 🙂

btw, I did not update the VCR tests as we've discussed that it'd be best
to remove the VCR for `text-to-video`. Maybe we should remove them here?
**EDIT**: removed the text-to-video tests in
[f8a6386](https://github.com/huggingface/huggingface.js/pull/1292/commits/f8a638657973ff8ce31b19581b50ed68d4c87b93).

I've tested it locally with
[tencent/HunyuanVideo](https://huggingface.co/tencent/HunyuanVideo) for
which the generation takes more than 2min and it works fine:


https://github.com/user-attachments/assets/3cd38900-c4ed-4b28-ae79-8a4e724f58d1

---------

Co-authored-by: Julien Chaumond <julien@huggingface.co>
Co-authored-by: Simon Brandeis <33657802+SBrandeis@users.noreply.github.com>