[Snippets] Fix text-to-image fetch snippets (#1317)
with @Wauplin, we realized that some `text-to-image` js snippets were
incorrect, the key use to pass the prompt is different from a provider
to another, so it's better to get the inputs from the body instead.

**EDIT**: this PR does not fix totally the issue, since the input keys
are customized per provider in the `textToImage` function. This is fixed
in #1315.

**EDIT 2**: in a later PR we will apply a linter on the generated
snippets which should make the snippets much nicer!