[Inference] Implement a "1 class = 1 provider<>task pair" logic to isolate provider-specific code (#1315)
The Goal of this PR is to isolate provider-specific logic into its own
class/file. Similar to what we have in the Python client, this gives a
clear boundary between generic task implementation and provider-specific
implementation details.

This makes updating or adding a new provider much easier, adding support
for a new provider will be as much as easier as:
1) register the provider in the mapping.
2) create a new `packages/inference/src/providers/{provider_name}.ts`.
3) implement the methods that require custom handling.

## Main changes
- Added a new `TaskProviderHelper` abstract class that defines a common
interface (`makeBody`, `preparePayload` `makeRoute`, `getResponse`,
`prepareHeaders`, etc.) that all provider-specific helpers must
implement.
- Added `PROVIDERS` registry that maps any provider<>task to the
corresponding `TaskProviderHelper` subclass that implements the
provider<>task logic.
- Added a lookup function (`getProviderHelper`) that takes a provider
and task and returns the appropriate TaskProviderHelper instance from
the `PROVIDERS` registry.
- Implemented the task helpers for providers:
     - [x] All task helpers have been implemented for all providers.

🙏 Any feedback is welcome as the implementation is quite biased towards
the Python but we might want to do some things differently here.

---------

Co-authored-by: SBrandeis <simon@huggingface.co>