Add CrossEncoder to the Sentence Transformers "How to Use" snippet & patch SetFit (#1337)
Hello!

## Pull Request overview
* Add CrossEncoder to the Sentence Transformers "How to Use" snippet
* Patch "How to Use" for SetFit models (that are also tagged as Sentence
Transformers)

## Details
This PR adds the `CrossEncoder` class to the "How to Use" snippets. I'm
working on feature parity of these models with the more common
`SentenceTransformer` models, and the recent Sentence Transformers v4.0
release was a big step in that.
We're already at 150 models:
https://huggingface.co/models?pipeline_tag=text-ranking&library=sentence-transformers

Beyond that, this ensures that `[widgetExample.source_sentence,
...widgetExample.sentences]` doesn't crash, apparently that currently
happens for some SetFit models, e.g.:
https://huggingface.co/faodl/setfit-paraphrase-mpnet-base-v2-5ClassesDesc-multilabel-augmented?library=sentence-transformers

- Tom Aarsen