gguf : add findNearestQuantType (#1421)
In this PR:
- Move `GGMLFileQuantizationType` to `tasks`
- Update the list of `GGMLFileQuantizationType` (NOTE: a **File** can
contains multiple quants, for example Q4_K_M **File** is Q4_K + Q6_K)
- For `GGMLQuantizationType`, add TQ1_0 and TQ2_0 tenary quants
- Add `findNearestQuantType` (see below)

## findNearestQuantType

This function is useful for `/v2` registry endpoint with **text + vision
models**, in case we want to pick the correspond vision model that can
be paired with a text model.

The main issue is that text model can go lower than Q4, like Q3/2/1, but
it is **not** the case for vision model, as vision model is quite
sensitive to quantization.

On @bartowski1182 repos , most of vision models will have BF16, F16 and
maybe Q8_0 versions. The idea is:
- If user pick BF16/F16/Q8_0 text model, we pair it to the correspond
BF16/F16/Q8_0
- If user pick something else, like Q4_K_M, we find the nearest quant to
be paired with. It's Q8_0 in this case