[Inference Providers] isolate `image-to-image` payload build for HF Inference API (#1439)
This PR is a prerequisite for #1427. 

It refactors the payload construction for `hf-inference` by isolating it
into a separate async function. Note that adding a new async function to
build the payload is necessary because
`HFInferenceImageToImageTask.preparePayload` cannot be made async, yet
the payload construction requires asynchronous operations. A similar
pattern has already been implemented for
[automaticSpeechRecognition](https://github.com/huggingface/huggingface.js/blob/361a0fad4c68943592f0bcbe41592d785eedcb81/packages/inference/src/tasks/audio/automaticSpeechRecognition.ts#L36)
with fal.