Harmonize inferenceProviderMapping additional parameter in modelInfo / listModel (#1515)
Equivalent to https://github.com/huggingface/huggingface_hub/pull/3022
in Python.

Main problem is that `expand[]=inferenceProviderMapping` do not return
the same data structure when getting model info or listing models. We
will fix this in ~3 months (?) but in the meantime we want a client
compatible with both current and future structure (see
https://github.com/huggingface/huggingface_hub/pull/3022 for more
details).


**TODO:**
- [x] adapt call in `@huggingface/inference`