Fix misleading error message while streaming (#1591)
In `innerRequest`/`innerStreamingRequest`, if an error occur we raise a
`InferenceClientProviderApiError` with a HTTP Request object in it. To
avoid bloating the logs, a `requestArgsToJson` helper is applied to
clean-up the body. However we do not pass the body to its but the
`requestArgs` instead i.e. **before** provider-specific logic has been
applied. This is can be really misleading since the body shown in the
error is not necessarily the one sent in the request, making it hard to
investigate.

I also took the opportunity to make sure `accessToken` is never printed
in the logs (it is currently the case as part of `requestArgs`).

(also added a console.log for auto-provider)