Add custom inference logger for @hf/inference (#1598)
The goal is to pass a custom logging tool (needed mainly by moon). 

I chose a global overwrite instead of an option like fetch to avoid too
many code changes (btw, there are some places where fetch is not
properly overwritten by the options).

I think the best approach would be to have a separate package for
logging to use the same module in the hf.js ecosystem