Basic Xorb creation + Add xet-core WASM bindings (#1616)
This doesn't handle dedup at all yet, will not be used as is even if I
merge it.

What it does:

- Write chunks into xorb
- Write chunk headers
- Handle compression
- Handle xorb size limit and xorb chunk count limit

Questions:

- ~~Are there xorb headers? I'm writing chunk headers before each chunk,
but is there a xorb header at the beginning of the xorb to write?~~ No,
thanks @assafvayner
- For now I only test lz4 compression length against uncompressed
length, I do not bother with bg4. Should I test both?
- Is the xet backend hardened against invalid chunks/xorbs?

cc @assafvayner @seanses @hoytak @sirahd @rajatarya for viz and if you
have answers to the questions :)

---------

Co-authored-by: Sylvestre Bouchot <Kakulukian@users.noreply.github.com>