Richer conversational snippet for AutoModel (#1611)
I can't remember if we already discussed about this. This is an attempt
to make the `AutoModel` snippet more self-contained, for the particular
but frequent case of conversational transformers models. The pipeline
version already got this treatment in #1434, but the `AutoModel` version
is currently too barebones in comparison, as shown in the screenshot
below.

In my opinion this could reduce the friction for users that want to
experiment with the models in Colab or Kaggle.
cc @Vaibhavs10 @ariG23498 

<img width="789" height="534" alt="Screenshot 2025-07-11 at 14 05 15"
src="https://github.com/user-attachments/assets/558837e3-ec56-4100-afa1-7a54d3d79305"
/>

---------

Co-authored-by: vb <vaibhavs10@gmail.com>