Add dev-container (#1702)
Helpful to provide an image when you want to run the codebase (including
installling deps & extensions) inside a dev container, to avoid the
codebase + extensions to impact the rest of your system.

When opening the project, VSCode & Cursor will now prompt you to reopen
the project in a dev container.

There is also no need anymore to "trust" the codebase through VSCode's
dialog