xet chunker is created for each file instead of reused (#1738)
We need to create a fresh Chunker object for each file we process to
avoid the last bytes from the previous file to impact the hash used to
generate chunk boundaries.

This gets ride of the `try {} finally {chunker.free()}` at the end of
file processing so this causes a lot of code to be de-indented, changes
are pretty minimal.

Co-authored-by: coyotte508 <coyotte508@protonmail.com>