Update chunkCache during backtracking after deduplication
Update the chunk indexes in the chunkCache after deduplication to match the new xorb chunks.
They were already updated in the passed `chunkMetadata` array, but not in the chunkCache. Which meant
that if the chunks were requested again eg (due to uploading another file with shared content in the same `uploadShards`)
the chunk indexes provided would be incorrect.