xet: Update chunk indexes in the chunkCache after deduplication (#1780)
Update the chunk indexes in the chunkCache after deduplication to match
the new xorb chunks.

The problem happens if a previous chunk in the xorb is removed during
backtrack deduplication. Then the chunk indexes for the following chunks
in teh xorb were updated in the passed `chunkMetadata` array, but not in
the chunkCache.

Which meant that if the chunks were requested again eg (due to uploading
another file with shared content in the same `uploadShards` call) the
chunk indexes provided would be incorrect.

cc @assafvayner @mishig25 for viz

Follow up #1771 and #1779

---

Also improved the debug/replay script to support multiple files upload,
and added a test