[Inference Snippets] Reorder clients: openai > huggingface > requests/fetch (#1789)
Quickly discussed with @hanouticelina. Order of definition of the
inference snippets was a bit random. Let's always show "openai" client
when possible (i.e. when `chat completion`), then
`huggingface_hub`/`huggingface.js` and then the raw `fetch`/`requests`.

By default on the Hub we are already showcasing `openai` first. This
update will align the snippets in the docs as well.