📈 Download tracking for new diff-interpretation-tuning project (#1791)
Registering a new library per the instructions on
https://huggingface.co/docs/hub/models-adding-libraries#register-your-library
and https://huggingface.co/docs/hub/models-download-stats.

This library is used by
https://huggingface.co/diff-interpretation-tuning/loras model repo, and
we are adding it so that downloads can be accurately counted. Note that
every .pt file contains one or more adapter files, hence we track the
total number of .pt file downloads.

---------

Co-authored-by: Lucain <lucainp@gmail.com>