Make safetensor shard filename parsing more flexible (#1817)
Context:

A lot of deepseek models have unconventional number of digits for the
total shard number in their filename, which leads to a filename parsing
fail. It breaks the file info navigation on the hub:



https://github.com/user-attachments/assets/75676c00-387c-4a92-808d-0d84caba7e51




[here](https://huggingface.co/deepseek-ai/DeepSeek-V3.2-Exp?show_file_info=model-00001-of-000163.safetensors)
for instance, `model-00001-of-000163.safetensors` should be
`model-00001-of-00163.safetensors` (there's 1 extra "0" in
"model-00001-of-.safetensors")

Most of deepseek models with a lot of shards have this issue, since they
are important models, I suggest we make the regex more flexible to
accept a variable number of digits for shards in the filename.

Would that be acceptable?