[Inference] Let's use lowercase left/right truncation direction parameter (#1841)
First spotted by @alvarobartt on
[slack](https://huggingface.slack.com/archives/C064Z3GRU3U/p1762871481747849?thread_ts=1743434022.621759&cid=C064Z3GRU3U)
(internal).

The `FeatureExtractionInputTruncationDirection` parameter defined in the
feature-extraction specs currently have capitalized values `Left` /
`Right`. This comes from TEI implementation from which we pulled the
specs. Let's switch it to lowercase values for consistency with all
other specs.

In addition to this PR, we should:
- update TEI to accept `left`/`right` (not even sure we want backward
compatibility on this) => @alvarobartt would you be able to handle this?
- update `huggingface_hub` => I can take care of this
https://github.com/huggingface/huggingface_hub/pull/3548
- update `hub-docs` => will happen automatically

I think it's fine to have breaking change for this. Moreover, specs /
docs are there more for information but are rarely enforced (expect in
TEI) so it shouldn't break anything client-side.