Add `image-text-to-image` and `image-text-to-video` tasks (#1866)
The goal of this new tasks is to support models that take in both image
and text input and output either image or video.

The goal of this PR is making the tasks as analogous to `image-to-image`
and `image-to-video` as possible, with the only difference that the
image input should now be **optional**, as an empty image and a valid
prompt should still work for a model like FLUX.2 (supports both
text-to-image and image-to-image tasks) or LTX Video (both text-to-video
and image-to-video)

Once this is in, I'll also have a widget PR in Moon to support this task
in the model cards / widgets etc. and a follow up PR adding this to the
inference providers, so that we can then PR repos to change the task for
compatible models

---------

Co-authored-by: Merve Noyan <merveenoyan@gmail.com>